refactor(UnexpectedItem)!: moved items into `VanillaGen` and flattened, with `Item` postfix on all names
